% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat_gff.R
\name{concat_gff}
\alias{concat_gff}
\title{Helper function. Takes in a GFF3 feature table and a contig position table created by `get_concat_position()`, returns an altered GFF3 feature table once all contigs are concatenated. Only "CDS" features are kept.}
\usage{
concat_gff(gff, contig)
}
\arguments{
\item{gff}{GFF3 annotation table}

\item{contig}{Contig position table, which is the output of `get_concat_position()`.}
}
\description{
Helper function. Takes in a GFF3 feature table and a contig position table created by `get_concat_position()`, returns an altered GFF3 feature table once all contigs are concatenated. Only "CDS" features are kept.
}
