% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat_vcf.R
\name{concat_vcf}
\alias{concat_vcf}
\title{Helper function. Takes in a vcfR object and a contig position table created by `get_concat_position()`, returns an altered vcf table once all contigs are concatenated.}
\usage{
concat_vcf(vcf, contig, remove_consensus = F, count = F)
}
\arguments{
\item{vcf}{vcfR object, ideally generated from FreeBayes.}

\item{contig}{Contig position table, which is the output of `get_concat_position()`.}

\item{remove_consensus}{if a mutation should be removed had all the genomes contained that same mutation. Defaults to FALSE and only works when count is TRUE.}

\item{count}{if the presence absence of the loci should be counted, defaults to FALSE.}
}
\description{
Helper function. Takes in a vcfR object and a contig position table created by `get_concat_position()`, returns an altered vcf table once all contigs are concatenated.
}
