% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw}
\alias{draw}
\title{Draw a heatmap of concatenated mutation loci.}
\usage{
draw(mut_table, inv = F, half = F, phylo = NULL)
}
\arguments{
\item{mut_table}{Output sheet from the main function `vcgff_gen()`.}

\item{inv}{If mutations that more than half of the genomes have should be inverted, defaults to FALSE}

\item{half}{If mutations that exactly half of the genomes have should be inverted, defaults to FALSE}

\item{phylo}{The vector of rownames or order of genomes that is wished to be drawn. You can provide a tree and use `tree$order` as this vector.}
}
\description{
Draw a heatmap of concatenated mutation loci.
}
