% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_SNP_phylo.R
\name{draw_SNP_phylo}
\alias{draw_SNP_phylo}
\title{Visualization function to draw SNP distributions by position in the form of sliding windows.}
\usage{
draw_SNP_phylo(
  mut_table,
  half = F,
  range = c(0, 2e+06),
  win = (range[2] - range[1])/200,
  GGTR = as.phylo(htree(mut_table)) \%>\% ggtree(ladderize = F) + geom_tiplab(align = T)
    + geom_rootpoint() + xlim_tree(5.5),
  Size = 7
)
}
\arguments{
\item{mut_table}{Output sheet from the main function `vcgff_gen()`.}

\item{half}{If mutations that exactly half of the genomes have should be inverted, defaults to FALSE}

\item{range}{The range of the DNA region to be plotted, defaults to c(0,2000000)}

\item{win}{size of the sliding window, defaults to range_length/200}

\item{GGTR}{A ggtree object of the genomes wished to be drawn. Defaults to `as.phylo(htree(mut_table)) %>% ggtree(ladderize = F)+geom_tiplab(align = T)+geom_rootpoint()+ xlim_tree(5.5)`.}

\item{Size}{Text size, defaults to 7}
}
\description{
Visualization function to draw SNP distributions by position in the form of sliding windows.
}
