% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst_.R
\name{fst_}
\alias{fst_}
\title{Calculates FST of all mutation loci, given an output sheet from the main function `vcgff_gen()` and a vector of suspected subpopulations.}
\usage{
fst_(mut_table, subpop_vector)
}
\arguments{
\item{mut_table}{Output sheet from the main function `vcgff_gen()`.}

\item{subpop_vector}{A vector that records the suspected subpopulation that would be examined for FST. The vector should be numerical, in the order of the column name that appeared in the mut_table.}
}
\description{
Calculates FST of all mutation loci, given an output sheet from the main function `vcgff_gen()` and a vector of suspected subpopulations.
}
