% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_concat_position.R
\name{get_concat_position}
\alias{get_concat_position}
\title{Helper function. Takes in a fragmented DNAString object, returns a data frame of contig positions once all contigs are concatenated.}
\usage{
get_concat_position(dna_contigs)
}
\arguments{
\item{dna_contigs}{fragmented DNAString object}
}
\description{
Helper function. Takes in a fragmented DNAString object, returns a data frame of contig positions once all contigs are concatenated.
}
