% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_extended_CDS.R
\name{get_extended_CDS}
\alias{get_extended_CDS}
\title{Helper function. Takes in the original DNA sequence, start and end positions of a gene, and the information of a mutation. Calculates the altered CDS sequence had the mutation taken place, considering frameshifts.}
\usage{
get_extended_CDS(dna_c, contig, start, end, POS, strand, seqid, REF, ALT)
}
\arguments{
\item{dna_c}{Concatenated DNAString}

\item{contig}{Contig position table, which is the output of `get_concat_position()`.}

\item{start}{Start position of a gene on the concatenated DNA.}

\item{end}{End position of a gene on the concatenated DNA.}

\item{POS}{Position of the mutation on the concatenated DNA}

\item{strand}{Strand direction of the gene}

\item{seqid}{If the original DNA sequence is fragmented, this is the contig number of the mutation on the fragmented DNAString object.}

\item{REF}{The sequence before the mutation (reference)}

\item{ALT}{The sequence after the mutation (query)}
}
\description{
Helper function. Takes in the original DNA sequence, start and end positions of a gene, and the information of a mutation. Calculates the altered CDS sequence had the mutation taken place, considering frameshifts.
}
