% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_count.R
\name{gt_count}
\alias{gt_count}
\title{Helper function. Takes in a vcfR object and returns its gt section into a 0/1 presence absence matrix.}
\usage{
gt_count(vcf, count = F)
}
\arguments{
\item{vcf}{vcfR object. Ideally a multi-vcf file generated from FreeBayes.}

\item{count}{A boolean that indicates if the presence absence matrix transformation would be done. Defaults to FALSE, which will only return the gt section without alteration.}
}
\description{
Helper function. Takes in a vcfR object and returns its gt section into a 0/1 presence absence matrix.
}
