% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htree.R
\name{htree}
\alias{htree}
\title{Create a hierarchical clustering phylo object using the output mutation sheet from the main function `vcgff_gen()`.}
\usage{
htree(mut_table, METHOD = "average")
}
\arguments{
\item{mut_table}{Output sheet from the main function `vcgff_gen()`.}

\item{METHOD}{method of hierarchical clustering, defaults to "average".}
}
\description{
Create a hierarchical clustering phylo object using the output mutation sheet from the main function `vcgff_gen()`.
}
