% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data_long.R
\name{input_data_long}
\alias{input_data_long}
\title{Inputs VCF and GFF data at the same time.}
\usage{
input_data_long(
  vcf_path,
  long_gff_path,
  dna_path = NA,
  vcf_name = "vcf",
  gff_name = "gff",
  concat_dna_name = "dna",
  dna_name = paste(concat_dna_name, "_contigs", sep = "")
)
}
\arguments{
\item{vcf_path}{Relative path to the VCF file}

\item{long_gff_path}{Relative path to the long GFF3 file}

\item{dna_path}{Relative path to the FASTA file, defaults to NA; if specified, will interpret the provided GFF3 as a short GFF file (FASTA component will be ignored).}

\item{vcf_name}{Variable name which stores the VCF file. Defaults to variable "vcf"in the global environment, as a vcfR object}

\item{gff_name}{Variable name which stores the GFF annotation table. Defaults to variable "gff" in the global environment, as a dataframe}

\item{concat_dna_name}{Variable name which stores the concatenated FASTA. Defaults to "dna" as a DNAString object}

\item{dna_name}{Variable name which stores the fragmented FASTA. Defaults to "dna_contigs" (or concat_dna_NAME+"_contigs") as a DNAString object}
}
\description{
Inputs VCF and GFF data at the same time.
}
