% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mut_pos_count.R
\name{mut_pos_count}
\alias{mut_pos_count}
\title{Helper function. Summarize mutation positions based on a sliding window. Returns a data frame.}
\usage{
mut_pos_count(mut_table, win = 10000, half = F, maxlen = 2e+06)
}
\arguments{
\item{mut_table}{Output sheet from the main function `vcgff_gen()`.}

\item{win}{Size of the sliding window, defaults to 10000}

\item{half}{If mutations that exactly half of the genomes have should be inverted, defaults to FALSE}

\item{maxlen}{The maximum length of the examined sequences.}
}
\description{
Helper function. Summarize mutation positions based on a sliding window. Returns a data frame.
}
