% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_long_gff3.R
\name{read_long_gff3}
\alias{read_long_gff3}
\title{Splits a long GFF3 file into two files and stores the annotation into the variable "gff_NAME" (which is `gff` by default) and FASTA into "concat_dna_NAME" (concatenated DNA, `dna` by default) and "dna_NAME" (DNA in contigs, `dna_contigs` by default).}
\usage{
read_long_gff3(
  longgff_path,
  keepDNA = T,
  gff_NAME = "gff",
  concat_dna_NAME = "dna",
  dna_NAME = paste(concat_dna_NAME, "_contigs", sep = "")
)
}
\arguments{
\item{longgff_path}{Relative path to the long GFF3 file}

\item{keepDNA}{Option to keep DNA. Defaults to TRUE.}

\item{gff_NAME}{Variable name which stores the GFF annotation table. Defaults to "gff"}

\item{concat_dna_NAME}{Variable name which stores the concatenated FASTA. Defaults to "dna"}

\item{dna_NAME}{Variable name which stores the fragmented FASTA. Defaults to "dna_contigs" (or concat_dna_NAME+"_contigs")}
}
\description{
Splits a long GFF3 file into two files and stores the annotation into the variable "gff_NAME" (which is `gff` by default) and FASTA into "concat_dna_NAME" (concatenated DNA, `dna` by default) and "dna_NAME" (DNA in contigs, `dna_contigs` by default).
}
