% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcgff_gen.R
\name{vcgff_gen}
\alias{vcgff_gen}
\title{Main function. Takes in VCF and GFF data, and outputs a dataframe that records all CDS-altering mutations, identifies their outcomes (non-synonymous, synonymous, frameshift etc.), and predicts the post-mutation translation product. Intergenic mutations will be ignored.}
\usage{
vcgff_gen(
  vcf = vcf,
  gff = gff,
  dna_c = dna,
  dnaContigs = dna_contigs,
  virulence = F,
  count = F,
  remove_consensus = F,
  fix_contig_name = NA
)
}
\arguments{
\item{vcf}{vcfR object, ideally generated from FreeBayes. Defaults to global environment object 'vcf'.}

\item{gff}{Reference sequence annotation GFF3 feature table. Ideally generated by Bakta. Defaults to global environment object 'gff'.}

\item{dna_c}{Reference concatenated DNA sequence, DNAString object. Defaults to global environment object 'dna'.}

\item{dnaContigs}{Reference fragmented DNA sequence, DNAString object. Defaults to global environment object 'dna_contigs'.}

\item{virulence}{BETA. Defaults to FALSE. If TRUE, marks all virulence genes, which is indicated by the description of the translation product in the gff file.}

\item{count}{if the presence absence of the loci should be counted, defaults to FALSE.}

\item{remove_consensus}{if a mutation should be removed had all the genomes contained that same mutation. Defaults to FALSE and only works when count is TRUE.}

\item{fix_contig_name}{BETA. Defaults to NA. If the contig names on the GFF file and the VCF file does not match, you can provide a vector of fixed contig names, which matches the names on the gff file. Contig names on the VCF and the dnaContigs will be changed accordingly.}
}
\description{
Main function. Takes in VCF and GFF data, and outputs a dataframe that records all CDS-altering mutations, identifies their outcomes (non-synonymous, synonymous, frameshift etc.), and predicts the post-mutation translation product. Intergenic mutations will be ignored.
}
